Remove ExtractorInterface and move DefinitionExtractor to the "Definition" folder
