+++ src/Definition/DefinitionExtractor.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Factory\Extractor;
+namespace Yiisoft\Factory\Definition;
 
 use Closure;
 use ReflectionClass;
@@ -11,18 +11,31 @@ use ReflectionFunctionAbstract;
 use ReflectionNamedType;
 use ReflectionParameter;
 use ReflectionUnionType;
-use Yiisoft\Factory\Definition\ClassDefinition;
-use Yiisoft\Factory\Definition\DefinitionInterface;
-use Yiisoft\Factory\Definition\ParameterDefinition;
 use Yiisoft\Factory\Exception\NotInstantiableException;
 
 /**
- * Class DefinitionExtractor
- * This implementation resolves dependencies by using class type hints.
+ * This class resolves dependencies by using class type hints.
  * Note that service names need not match the parameter names, parameter names are ignored
+ *
+ * @internal
  */
-class DefinitionExtractor implements ExtractorInterface
+final class DefinitionExtractor
 {
+    private static ?self $instance = null;
+
+    private function __construct()
+    {
+    }
+
+    public static function getInstance(): self
+    {
+        if (self::$instance === null) {
+            self::$instance = new self();
+        }
+
+        return self::$instance;
+    }
+
     /**
      * @psalm-param class-string $class
      *
